% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmClassif.BIC.R
\name{gmmClassif.BIC}
\alias{gmmClassif.BIC}
\title{Critère BIC}
\usage{
gmmClassif.BIC(logML, v, n)
}
\arguments{
\item{logML}{La log vraissemblance}

\item{v}{Le nombre de variable du modèle}

\item{n}{Le nombre d'observation}
}
\value{
BIC
}
\description{
Critère BIC
}
\examples{
require(rsample)

# Données
split <- initial_split(iris, prop=.70, strata="Species")

iris_train = training(split)
iris_test = testing(split)

X_train = iris_train[,1:4]
y_train = iris_train[, 5]
X_test = iris_test[, 1:4]
y_test = iris_test[, 5]

# Variables descriptives
X = X_train

# Variable reponse
z = y_train

model <- gmmClassif.EM(X = X, z = z, K = 3)
gmmClassif.BIC(model$logML, ncol(X), nrow(X))
}
