% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmClassif.EM.R
\name{gmmClassif.EM}
\alias{gmmClassif.EM}
\title{Algorithme Expectation Maximization}
\usage{
gmmClassif.EM(X, z, K, eps = 10^-3)
}
\arguments{
\item{X}{Jeu de données (variables descriptives)}

\item{z}{Variable reponse}

\item{K}{Le nombre de cluster}

\item{eps}{Le critère de convergence}
}
\value{
Nos paramètres maximisé ainsi que sa log-vraissemblance associé
}
\description{
Algorithme Expectation Maximization
}
\examples{
require(rsample)

# Données
split <- initial_split(iris, prop=.70, strata="Species")

iris_train = training(split)
iris_test = testing(split)

X_train = iris_train[,1:4]
y_train = iris_train[, 5]
X_test = iris_test[, 1:4]
y_test = iris_test[, 5]

# Variables descriptives
X = X_train

# Variable reponse
z = y_train

gmmClassif.EM(X = X, z = z, K = 3)
}
