% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmClassif.logML.R
\name{gmmClassif.logML}
\alias{gmmClassif.logML}
\title{Log-Vraissemblance du Modèle de Mélange}
\usage{
gmmClassif.logML(X, model)
}
\arguments{
\item{X}{Les données}

\item{model}{Les paramètres du modèle}
}
\value{
La log-vraissemblance
}
\description{
Log-Vraissemblance du Modèle de Mélange
}
\examples{
require(rsample)

# Données
split <- initial_split(iris, prop=.70, strata="Species")

iris_train = training(split)
iris_test = testing(split)

X_train = iris_train[,1:4]
y_train = iris_train[, 5]
X_test = iris_test[, 1:4]
y_test = iris_test[, 5]

# Variables descriptives
X = X_train

# Variable reponse
z = y_train

model = gmmClassif.MLParameter(X = X, z = z, K = 3)

gmmClassif.logML(X = X, model = model)
}
