% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmClassif.MLParameter.R
\name{gmmClassif.MLParameter}
\alias{gmmClassif.MLParameter}
\title{Estimation des paramètres par estimation du maximum de vraissemblance}
\usage{
gmmClassif.MLParameter(X, z, K, tik = NA, method = "ML")
}
\arguments{
\item{z}{Variable réponse ou à prédire}

\item{K}{Le nombre de cluster}

\item{tik}{La probabilité que les xi appartienne au cluster k}

\item{method}{La méthode utilisé Vraissemblance "classique" (ML) ou vraissemblance pondérée}

\item{x}{Données}
}
\value{
Paramètre proportion
}
\description{
Estimation des paramètres par estimation du maximum de vraissemblance
}
\examples{
require(rsample)

# Données
split <- initial_split(iris, prop=.70, strata="Species")

iris_train = training(split)
iris_test = testing(split)

X_train = iris_train[,1:4]
y_train = iris_train[, 5]
X_test = iris_test[, 1:4]
y_test = iris_test[, 5]

# Variables descriptives
X = X_train

# Variable reponse
z = y_train

model = gmmClassif.MLParameter(X = X, z = z, K = 3)
}
