% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmClassif.predict.R
\name{gmmClassif.predict}
\alias{gmmClassif.predict}
\title{Prédiction du cluster et calcul de la probabilité d'appartenance}
\usage{
gmmClassif.predict(X, model)
}
\arguments{
\item{X}{Les données}

\item{model}{les paramètres du modèle}
}
\value{
La probabilité et la prédiction que les x_i appartiennent au cluster k
}
\description{
Prédiction du cluster et calcul de la probabilité d'appartenance
}
\examples{
require(rsample)

# Données
split <- initial_split(iris, prop=.70, strata="Species")

iris_train = training(split)
iris_test = testing(split)

X_train = iris_train[,1:4]
y_train = iris_train[, 5]
X_test = iris_test[, 1:4]
y_test = iris_test[, 5]

# Variables descriptives
X = X_train

# Variable reponse
z = y_train

model = gmmClassif.MLParameter(X = X, z = z, K = 3)

pred = gmmClassif.predict(X = X_test, model = model)

# Erreur de classification : Matrice de Confusion
table(pred$z, y_test)
}
