% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmClassif.R
\name{gmmClassif}
\alias{gmmClassif}
\title{Modèle de Mélange Gaussien (GMM) en classification
Fonction de Classification}
\usage{
gmmClassif(Xtrain, Xtest, z, K, eps = 10^-3)
}
\arguments{
\item{Xtrain}{Variables descriptives des individus de l'échantillon d'apprentissage}

\item{Xtest}{Variables descriptives des individus de l'échantillon test}

\item{z}{Variable réponse à prédire pour les individus de l'échantillon d'apprentissage}

\item{K}{Le nombre de cluster}

\item{eps}{Le critère de convergence de l'algo EM}
}
\value{
La variable réponse prédite dans le cluster k
Le nom du modèle
Les paramètres du modèles ainsi que le nombre de cluster, la proportion de mélange
}
\description{
Modèle de Mélange Gaussien (GMM) en classification
Fonction de Classification
}
\examples{
require(rsample)

# Données
split <- initial_split(iris, prop=.70, strata="Species")

iris_train = training(split)
iris_test = testing(split)

X_train = iris_train[,1:4]
y_train = iris_train[, 5]
X_test = iris_test[, 1:4]
y_test = iris_test[, 5]

gmm = gmmClassif(Xtrain = X_train, Xtest = X_test, z = y_train, K = 3)

# Affichage graphique
require(cluster)
clusplot(X_test, gmm$pred$z, lines=0,
         shade=TRUE, color=TRUE,
         labels=2, plotchar=FALSE,
         span=TRUE, main=paste("Clusters of Iris"), dimens=c(1, 2, 3))
}
